requirements.txt
transformers
datasets
trl
evaluate
peft
hydra-core
rich
pandas
seaborn
rootutils
autorootcwd
wandb
bitsandbytes
accelerate
langdetect
scikit-learn
optimum
auto-gptq